added test coverage
