+++ src/test/java/uk/gov/hmcts/sptribs/caseworker/event/CaseworkerEditCaseTest.java
@@ -1,0 +1,36 @@
+package uk.gov.hmcts.sptribs.caseworker.event;
+
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.junit.jupiter.MockitoExtension;
+import uk.gov.hmcts.ccd.sdk.ConfigBuilderImpl;
+import uk.gov.hmcts.ccd.sdk.api.Event;
+import uk.gov.hmcts.sptribs.ciccase.model.CaseData;
+import uk.gov.hmcts.sptribs.ciccase.model.State;
+import uk.gov.hmcts.sptribs.ciccase.model.UserRole;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static uk.gov.hmcts.sptribs.caseworker.event.CaseworkerEditCase.CASEWORKER_EDIT_CASE;
+import static uk.gov.hmcts.sptribs.testutil.ConfigTestUtil.createCaseDataConfigBuilder;
+import static uk.gov.hmcts.sptribs.testutil.ConfigTestUtil.getEventsFrom;
+
+@ExtendWith(MockitoExtension.class)
+public class CaseworkerEditCaseTest {
+    @InjectMocks
+    private CaseworkerEditCase caseworkerEditCase;
+
+    @Test
+    void shouldAddConfigurationToConfigBuilder() {
+        //given
+        final ConfigBuilderImpl<CaseData, State, UserRole> configBuilder = createCaseDataConfigBuilder();
+
+        //when
+        caseworkerEditCase.configure(configBuilder);
+
+        //then
+        assertThat(getEventsFrom(configBuilder).values())
+            .extracting(Event::getId)
+            .contains(CASEWORKER_EDIT_CASE);
+    }
+}
